requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.29
marshmallow==3.21.1
pytest==7.4.4
python-dotenv==1.0.1
Werkzeug==2.3.7 